Final edits on question text
I removed an entire sentence, and struggled with whether or not this was a good idea. My thinking is that the information about the configuration folder was confusing, but not in a way that provided a question distractor. I wanted the question to be clear, and for the learner to have to think about the command or syntax. I also considered how I would read this sentence if English were not my first language and how I would parse this information. In the end, I decided to go with the shortest, clearest question possible. But I definitely concede that my lack of experience with Terraform could mean that I was removing pertinent information. If this were real life, I would ask my SME a few questions:
1. What exactly are we testing with this question? I would want clarity on the knowledge and behavior we are looking to test. My thought was that the question is asking if the learner knows how to access the key, or file, using the correct Terraform syntax. Maybe this is not quite it, though.
2. Does the learner need the additional context in order to make this decision? Perhaps the location of the file is required to choose between these options. I looked at docs, but never found anything to make me feel 100% confident in my decision.